+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Enchancement: Improve EventProcessor nullability annotations (#1229).
 * Fix: Disable Gson HTML escaping
+* Enchancement: Support @SentrySpan and @SentryTransaction on classes and interfaces. (#1243)
 
 # 4.1.0